[backend/frontend] feat(injec-status): enhance Inject Result Clarity & Actionability (#5412)
